move specs next to implementation
