+++ pupil-api/src/controllers/auth.controller.spec.ts
@@ -2,10 +2,10 @@
 
 /* global describe spyOn */
 
-import { default as authController } from '../../controllers/auth.controller'
+import { default as authController } from './auth.controller'
 import * as httpMocks from 'node-mocks-http'
-import logger from '../../services/log.service'
-import { pupilAuthenticationService } from '../../services/pupil-auth.service'
+import logger from '../services/log.service'
+import { pupilAuthenticationService } from '../services/pupil-auth.service'
 
 describe('auth controller', () => {
   describe('route /auth', () => {
